Extract command session shell script
